+++ src/constants.ts
@@ -1,4 +1,4 @@
-export const VERSION = "1.4.1";
+export const VERSION = "1.5.0";
 
 export const Status = {
   QUEUED: "queued",